+++ packages/blip/test/unit/components/chart/stats.test.js
@@ -156,9 +156,8 @@ describe('Stats', () => {
         expect(wrapper.find({ 'data-testid': 'stat-averageGlucose' })).to.have.length(1)
         expect(wrapper.find({ 'data-testid': 'stat-standardDev' })).to.have.length(1)
         expect(wrapper.find({ 'data-testid': 'stat-totalInsulin' })).to.have.length(1)
-        _.forEach(expectedStats, statId => {
-          expect(wrapper.find(`#Stat--${statId}`)).to.have.length(1)
-        })
+        expect(wrapper.find({ 'data-testid': 'stat-carbs' })).to.have.length(1)
+        expect(wrapper.find('#Stat--coefficientOfVariation')).to.have.length(1)
       })
 
       it('should show all expected stats when bgSource prop is `smbg`', () => {
@@ -173,7 +172,7 @@ describe('Stats', () => {
         expect(wrapper.find({ 'data-testid': 'cbg-percentage-stats-readingsInRange' })).to.have.length(1)
         expect(wrapper.find({ 'data-testid': 'stat-averageGlucose' })).to.have.length(1)
         expect(wrapper.find({ 'data-testid': 'stat-totalInsulin' })).to.have.length(1)
-        expect(wrapper.find('#Stat--carbs')).to.have.length(1)
+        expect(wrapper.find({ 'data-testid': 'stat-carbs' })).to.have.length(1)
       })
 
       it('should render the Time in Auto stat for automated basal devices', () => {
@@ -306,16 +305,10 @@ describe('Stats', () => {
 
         expect(wrapper.find('.Stats').children()).to.have.length(4)
 
-        const expectedStats = [
-          'averageDailyDose',
-          'carbs'
-        ]
-
         expect(wrapper.find({ 'data-testid': 'cbg-percentage-stats-timeInRange' })).to.have.length(1)
         expect(wrapper.find({ 'data-testid': 'stat-averageGlucose' })).to.have.length(1)
-        expectedStats.forEach(statId => {
-          expect(wrapper.find(`#Stat--${statId}`)).to.have.length(1)
-        })
+        expect(wrapper.find({ 'data-testid': 'stat-carbs' })).to.have.length(1)
+        expect(wrapper.find('#Stat--averageDailyDose')).to.have.length(1)
       })
 
       it('should render the Time in Auto stat for automated basal devices', () => {
@@ -333,12 +326,12 @@ describe('Stats', () => {
         expect(wrapper.find('.Stats').children()).to.have.length(5)
         const expectedStats = [
           'averageDailyDose',
-          'timeInAuto',
-          'carbs'
+          'timeInAuto'
         ]
 
         expect(wrapper.find({ 'data-testid': 'stat-timeInRange' })).to.have.length(1)
         expect(wrapper.find({ 'data-testid': 'stat-averageGlucose' })).to.have.length(1)
+        expect(wrapper.find({ 'data-testid': 'stat-carbs' })).to.have.length(1)
         expectedStats.forEach(statId => {
           expect(wrapper.find(`#Stat--${statId}`)).to.have.length(1)
         })
@@ -353,16 +346,10 @@ describe('Stats', () => {
         wrapper.update()
         expect(wrapper.find('.Stats').children()).to.have.length(4)
 
-        const expectedStats = [
-          'averageDailyDose',
-          'carbs'
-        ]
-
         expect(wrapper.find({ 'data-testid': 'cbg-percentage-stats-readingsInRange' })).to.have.length(1)
         expect(wrapper.find({ 'data-testid': 'stat-averageGlucose' })).to.have.length(1)
-        expectedStats.forEach(statId => {
-          expect(wrapper.find(`#Stat--${statId}`)).to.have.length(1)
-        })
+        expect(wrapper.find({ 'data-testid': 'stat-carbs' })).to.have.length(1)
+        expect(wrapper.find('#Stat--averageDailyDose')).to.have.length(1)
       })
     })
   })