+++ packages/dumb/src/index.ts
@@ -34,6 +34,7 @@ import {
 } from './components/stats/cbg-standard-deviation/cbg-standard-deviation'
 import { BgClasses, CBGStatType } from './components/stats/models'
 import { TotalInsulinStatMemoized as TotalInsulinStat } from './components/stats/total-insulin/total-insulin-stat'
+import { TotalCarbsStatMemoized as TotalCarbsStat } from './components/stats/total-carbs/total-carbs-stat'
 import { ReservoirTooltip } from './components/tooltips/reservoir-tooltip/reservoir-tooltip'
 import { StatTooltip } from './components/tooltips/stat-tooltip/stat-tooltip'
 import Tooltip from './components/tooltips/tooltip/tooltip'
@@ -53,5 +54,6 @@ export {
   StatTooltip,
   TerminalTable,
   Tooltip,
-  TotalInsulinStat
+  TotalInsulinStat,
+  TotalCarbsStat
 }